+++ controller/main.py
@@ -1,12 +1,12 @@
-from typing import List
 import uvicorn
 from fastapi import FastAPI
 
-from controller.adapters import PersonAdapter, MatcherAdapter
+from matcher.application.adapter import MatcherAdapter
 from controller.exception_handlers import person_not_exists_exception_handler
 
-from person.domain.driven.ports import MalformedRequestException
-from person.domain.model import PersonDto, PersonOperationResponseDto
+from person.public.exception import MalformedRequestException
+from person.public.entities import PersonDto, PersonOperationResponseDto
+from person.public.driver.factory import get_manager as create_person_manager
 
 app = FastAPI()
 app.add_exception_handler(MalformedRequestException, person_not_exists_exception_handler)
@@ -14,13 +14,12 @@ app.add_exception_handler(MalformedRequestException, person_not_exists_exception
 
 @app.post("/person/save/")
 def save_person(request: PersonDto) -> PersonOperationResponseDto:
-    result = PersonAdapter().handle_save(request)
+    result = create_person_manager().handle_save(request)
     return result
 
 
 @app.post("/matcher/match_siblings/")
-def match_person_siblings(request: PersonDto) -> List[PersonDto]:
-
+def match_person_siblings(request: PersonDto) -> list[PersonDto]:
     return MatcherAdapter().handle_match_siblings(request)
