+++ matcher/application/adapter.py
@@ -1,17 +1,14 @@
 from typing import List
 
 from matcher.application.use_cases import MatchSiblingsUseCase
-from matcher.domain.driver.ports import MatcherManager
 from matcher.infrastructure.factory import create_mongo_matcher_manager
-from person.domain.model import PersonDto, person_from_dto, person_to_dto
+from matcher.public.driver.ports import MatcherManager
+from person.application.adapter import person_from_dto, person_to_dto
 from person.infrastructure.factory import create_mongo_person_repository
+from person.public.entities import PersonDto
 
 
-def get_manager() -> MatcherManager:
-    return DefaultManager()
-
-
-class DefaultManager(MatcherManager):
+class MatcherAdapter(MatcherManager):
     def handle_match_siblings(self, rq: PersonDto) -> List[PersonDto]:
         person = person_from_dto(rq)
         matches = MatchSiblingsUseCase(
