+++ matcher/application/driven/ports.py
@@ -1,8 +1,5 @@
 from abc import ABC, abstractmethod
 
-from person.domain.driven.ports import PersonRepository
-from person.domain.model import Person
-
 
 class MatcherManager(ABC):
     @abstractmethod
