+++ matcher/tests/test_application.py
@@ -1,17 +1,20 @@
 from unittest import TestCase
-from unittest.mock import MagicMock, Mock
+from unittest.mock import MagicMock
 
 from matcher.application.use_cases import MatchSiblingsUseCase
 from matcher.infrastructure.mongo_matcher import MongoMatcher
+from person.infrastructure.mongo_repository import MongoPersonRepository
 from settings.test import CLOUD_MONGO_CLIENT
-from person.domain.driven.ports import MalformedRequestException
+from person.public.exception import MalformedRequestException
 from person.domain.model import Person
 
 
 class MatchSiblingsUseCaseTestCase(TestCase):
     def setUp(self) -> None:
+        self._repository = MagicMock(MongoPersonRepository)
         self.manager = MagicMock(MongoMatcher(CLOUD_MONGO_CLIENT))
-        self.sut = MatchSiblingsUseCase(manager=self.manager)
+        self.sut = MatchSiblingsUseCase(manager=self.manager,
+                                        repository=self._repository)
 
     def test_match_siblings_unregistered_person(self):
         with self.assertRaises(MalformedRequestException) as context:
