+++ person/infrastructure/mongo_repository.py
@@ -1,6 +1,6 @@
 from typing import Optional
 from person.domain.model import Person
-from person.domain.driven.ports import PersonRepository
+from person.application.driven.ports import PersonRepository
 
 
 class MongoPersonRepository(PersonRepository):
