+++ person/public/driver/factory.py
@@ -1,0 +1,6 @@
+from person.public.driver.ports import PersonManager
+from person.application.adapter import PersonAdapter
+
+
+def get_manager() -> PersonManager:
+    return PersonAdapter()
