+++ person/public/exception.py
@@ -1,0 +1,2 @@
+class MalformedRequestException(Exception):
+    ...
